requirements.txt
Flask
Gunicorn
python-dotenv
pytest
psycopg2-binary
google-genai